refactor: moving makerQuerySearch function to separate folder - assistants
